Roast-a-Bot
this is a roast bot which roasts you whenever you type in a question, a project made to refine our python skills and make a fun project at the same time. it can be modified with Arduino and AI to increase its output sentences and increase its vocubalry and use question appropriate responcesand merge it with text to speech so that the bot can actually talk "trash".